refactor(structure): Standardize file naming and refactor NotFoundPage styles
- Renames component files to `index.tsx` for consistent imports.
- Moves inline styles from NotFoundPage to the main CSS file.
