+++ src/components/offer-card/index.tsx
