+++ src/pages/favourites-page/index.tsx
