--- structure.md
@@ -1,75 +1,0 @@
-```
-2670807-SIX-CITIES-6
-├── .github/
-│   └── workflows/
-│       ├── build.yml
-│       ├── check.yml
-│       └── publish.yml
-│
-├── markup/
-│   ├── css/
-│   │   ├── main.css
-│   │   └── main.css.map
-│   ├── fonts/
-│   ├── img/
-│   ├── favorites-empty.html
-│   ├── favorites.html
-│   ├── index.html
-│   ├── login.html
-│   ├── main-empty.html
-│   ├── main.html
-│   ├── offer-not-logged.html
-│   ├── offer.html
-│   ├── property-not-logged.html
-│   ├── property.html
-│   └── Readme.md
-│
-├── node_modules/
-│
-├── public/
-│   ├── css/
-│   │   ├── main.css
-│   │   └── main.css.map
-│   ├── fonts/
-│   └── img/
-│
-├── src/
-│   ├── app/
-│   │   ├── app-router.tsx
-│   │   ├── app.tsx             
-│   │   └── routes.ts 
-│   │
-│   ├── components/
-│   |   ├── offer-card/
-│   │   │   └── offer-card.tsx
-│   │   └── private-route/
-│   │       └── private-route.tsx
-│   │
-│   ├── pages/
-│   |   ├── favorites-page/
-│   │   |   └── favourites-page.tsx
-│   |   ├── login-page/
-│   │   |   └── login-page.tsx
-│   |   ├── main-page/
-│   │   |   └── main-page.tsx
-│   │   ├── not-found-page/
-│   │   │   └── not-found-page.tsx
-│   |   └── offer-page/
-│   │       └── offer-page.tsx
-│   │
-│   ├── index.tsx
-│   └── setupTests.ts
-│
-├── .editorconfig
-├── .eslintrc.cjs
-├── .gitattributes
-├── .gitignore
-├── Contributing.md
-├── index.html
-├── package-lock.json
-├── package.json
-├── Readme.md
-├── tsconfig.json
-├── tsconfig.node.json
-└── vite.config.ts
-```
